Organize code a bit

+++ src/main/java/il/org/spartan/utils/file.java
@@ -1,6 +1,7 @@
 package il.org.spartan.utils;
 
 import java.io.*;
+import java.nio.charset.*;
 import java.nio.file.*;
 
 import org.jetbrains.annotations.*;
@@ -12,12 +13,13 @@ import fluent.ly.*;
  * @since Dec 4, 2016 */
 public enum file {
   ;
-  private static void delete(final @NotNull String path) {
-    if (file.exists(path))
-      new File(path).delete();
+  private static void delete(final @Nullable String path) {
+    if (path != null && file.exists(path))
+        new File(path).delete();
   }
 
   private static void rename(final @NotNull String from, final @NotNull String to) {
+    
     file.delete(to);
     final Path source = Paths.get(from);
     try {
@@ -27,12 +29,33 @@ public enum file {
     }
   }
 
-  private static boolean exists(final @NotNull String path) {
+  private static boolean exists(final @Nullable String path) {
+    if (path == null)
+      return false;
     final File $ = new File(path);
     return $.exists() && !$.isDirectory();
   }
 
   public static void renameToCSV(final @NotNull String old) {
     file.rename(old, old + ".csv");
+  }
+
+  @NotNull public static String read(final @NotNull File f) throws IOException {
+    final @NotNull String ls = de.fault(System.getProperty("line.separator")).to("/");
+    System.err.println(ls.compareTo("\n"));
+    final @NotNull StringBuilder $ = new StringBuilder();
+    try (@NotNull BufferedReader reader = new BufferedReader(new FileReader(f))) {
+      for (String line = reader.readLine(); line != null; line = reader.readLine())
+        $.append(line).append(ls);
+    }
+    return $ + "";
+  }
+
+  @NotNull static String read(final @NotNull Path ¢) throws IOException {
+    return new String(Files.readAllBytes(¢), StandardCharsets.UTF_8);
+  }
+
+  @SuppressWarnings("null") @NotNull public static String read(final @NotNull String fileName) throws IOException {
+    return read(Paths.get(fileName));
   }
 }
